Make pullq a self-serving server
Instead of relying on external systems to serve the static ClojureScript
application and perform a period sync with the github data, this branch
adds the capacity directly to pullq by making it depend on aleph and
bidi to perform the serving.

The /data.edn URL functions as a single API endpoint for the
ClojureScript client, emulating the previous behavior by keeping the
state in an (atom) in-memory.

A simple thread is run periodically to refresh the in-memory state.